+++ src/main/java/com/jit/proxiBanqueV4/metier/IConseillerMetier.java
@@ -5,17 +5,13 @@ import java.util.List;
 
 import com.jit.proxiBanqueV4.entites.Client;
 import com.jit.proxiBanqueV4.entites.Conseiller;
-/**
- * 
- * @author Habachi,Cadi,Bourkha,Sid-Elkhir,Nouri
- *
- */
+
 public interface IConseillerMetier {
 	public Conseiller saveConseiller(Conseiller conseiller);
 	public List<Conseiller> listeConseillers();
-	public Conseiller getConseiller(int idConseiller);
+	public Conseiller getConseiller(Long idConseiller);
 	public boolean updateConseiller(Conseiller conseiller);
-	public boolean deleteConseiller(int idConseiller);
+	public boolean deleteConseiller(Long idConseiller);
 	public int seConnecter(String emailConseiller,String password);
 	
 	
@@ -23,9 +19,8 @@ public interface IConseillerMetier {
 	public Client createClient(Client client);
 	public Client getClient(Long idClient);
 	public boolean updateClient(Client client);
-	public List<Client> listeClients(int idConseiller);
-	public List<Client> alertDecouvert(int idConseiller);
-	
+	public List<Client> listeClients(Long idConseiller);
+	public List<Client> alertDecouvert(Long idConseiller);
 	
 
 }
